backend/requirements.txt
asgiref==3.8.1
configparser==7.1.0
Django==4.2.16
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
